Hide system properties from JMX
If there are any system properties that are replaced with values from
the vault then a role requirement `sensitive` is placed on the
java.lang:type=Runtime getSystemProperties access via JMX.

This is a bit of an overkill, to limit access to all system properties,
but I've found no other way of doing this. Will investigate other means.