Use basepom, unify Pax Exam tests, test robustness
Making use of basepom showed a few robustness issues in integration
tests as it was using random test execution order. In order to fix those
and reduce test time several tests were merged into one, this reduces
the number of test containers being started/stopped.

In order for the tests to be reliable in this scenario the
`vault-karaf-core` bundle needed to be restarted at the beginning of the
test, which showed that the `SecurityVault` instance has been held in
the `SecurityVaultFactory` singleton, so `SecurityVault` instance would
remain between the tests. This is of course not wanted and for runtime,
as would lead to unexpected behaviour.

So now Activator on `stop` tries to revert system properties to their
original values, not to leave them in the clear, `start`-ing the bundle
will replace them, as is expected.